Started working further on member notes
Now able to add notes to an enrollment in show view. Also possible in index view, but notes are not properly fetched from database yet. Still need to add an update method to the enrollments, as to change the notes, one needs to unenroll and enroll again. This is unwanted behaviour.