Re-add sidebar menu to member-view
As discussed yesterday, the menu and group-list is not required to be on
every page for the members, and we expect that the average member is
able to remember their own name. Because of this, the widget containing
the Gravatar, full name and email has been removed. The list of links
that followed the widget has been preserved in the mobile view, to
immediately present a menu to the user. A link to jump to the financial
overview will be added later, to decrease the chance a user will miss
the overview of activities on the homepage.