Fix a styling error occurring in Firefox
See #203

When as admin looking at a member with a checkout card in Firefox, the
payment method selection would overflow to the next line. This wouldn't
occur in Chrome/Chromium. After experimenting, I found out that
containing the select input in another div would fix this problem
without any regressions noticeable to me. But why this works I don't
know.

To make this fix work while retaining the same styling for this select
input, I generalized the background color and height scss entry for any
direct childs of a div#credit.input-group to any childs (so also further
descendants) of div#credit.input-group. As far as I know this doesn't
cause any changes apart from this intended change, since there is only
one element with the id credit and the class input-group, and it doesn't
have any other childs which this affects adversely.