Fix Activity.ended?
Problem: When the end_time of an activity is nil, the current check sets
the activity's end to midnight on end_date.

This caused the 'activity already ended'-bug for members.

The solution is still not ideal (as the activity will still be marked
as finished if there is no start_time), but should work better and allow
enrolling on the same day if there is no end_time.