Send emails in development in test mode (if token present)
This is not ideal: Rails supports a better test mode out of the box (by
writing to a file in tmp and/or dumping the mail to stdout), but this
method does not work because ApplicationMailer#mail is overridden.

The solution for this would be to use a proper Mailgun gem (so we can
switch in config/environments which method to use), but I'm not doing
that right now.

Sent emails will show up in the Mailgun logs.