Properly handle Participant enrollment conflict
Adds proper handling for a race condition where two clients attempt to
enroll a member. The response code will still be 409, but the admin's
response will return the expected information (which is still not
handled in the javascript), and the member's response will include an
error message.

Fixes #329, Sentry KOALA-PRODUCTION-F, Sentry KOALA-PRODUCTION-2