Fix regression bug in showing Checkout info
In a03433a I changed when the Checkout table is shown in the member show
view for admins. But now the table wasn't shown when a member had added
a card but no transactions yet, and an admin wanted to add a top-up
there. So I added the previously existing condition back as an OR.