admin: Hide pagination footer when searching for Members
Fixes #489.

The board has indicated that the search page's pagination is almost
never used, as the search is usually accurate enough to return the
requested Member at the top of the list, and only doesn't work if the
search query is not accurate enough (which can be corrected).

Removing the pagination footer is necessary because will_paginate only
works with result sets returned by Rails, while Fuzzily returns an Array
(as it does the pagination itself).

Also fixes an issue that caused the pagination numbers to render inside
the column for the limit dropdown.